Added marker publisher to mocap_node.cpp, I cheated and did not do it properly but needed to test. System crashed. unsure if it works.

+++ src/path_planning/src/ServiceCall.h
@@ -21,7 +21,7 @@ public:
     ServiceCall();
     lcar_msgs::WorldMap srv;
     ServiceCall(const ServiceCall& orig);
-    std::vector<int> call();
+    std::vector<double> call();
     virtual ~ServiceCall();
 private:
